rag-system/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.23.5 
huggingface-hub==0.14.1
transformers>=4.30.0
torch>=2.0.0
openai==0.28.0
typing-extensions==4.7.0
requests==2.28.2
tqdm==4.64.1
